Combat-patching magnetism into a better version of itself
Magnetism relied on global x/y points to snap to - this is way underperf because we need to recompute all of those each time the grid moves. This now relies on the EaselJS objects to be able to do the point coords conversion at comparison time, which necessitates a whole lot less work than listening to events and emptying/recreating the whole thousands of snap points.